Added debian package creation and fixed issues with post install script

+++ pkg/files/after_remove.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-userdel -r pine
+ret=true
+getent passwd pine >/dev/null 2>&1 && ret=false
+if $ret; then
+	userdel -r pine
+fi
